messageHandler/local_requirements.txt
boto3
requests
pymongo
python-dotenv
